refactor: ssr 방식으로 변경
