+++ src/feature/notice/NoticeTopNavigation.tsx
@@ -1,0 +1,10 @@
+'use client';
+
+import TopNavigation from '@/common/components/TopNavigation';
+import { postMessageToWebView } from '@/utils/webview';
+
+function NoticeTopNavigation() {
+  return <TopNavigation onClickBack={() => postMessageToWebView({ route: 'goBack' })} />;
+}
+
+export default NoticeTopNavigation;
